--- config.js
@@ -1,4 +1,0 @@
-const TELEGRAM_BOT_TOKEN = process.env.TELEGRAM_BOT_TOKEN;
-let API_KEYS = [];
-
-module.exports = { TELEGRAM_BOT_TOKEN, API_KEYS };
