+++ config/config.js
@@ -1,0 +1,4 @@
+const TELEGRAM_BOT_TOKEN = process.env.TELEGRAM_BOT_TOKEN;
+const BOTNAME = '@DegenTraderV2_bot';
+
+module.exports = { TELEGRAM_BOT_TOKEN, BOTNAME };
